refactor(tmux): split cleanup of captured buffer from capture
