+++ tmux-lib/src/lib.rs
@@ -41,7 +41,7 @@ pub(crate) mod parse;
 pub mod server;
 pub mod session;
 pub mod session_id;
-pub(crate) mod utils;
+pub mod utils;
 pub mod window;
 pub mod window_id;
